BasePowerTool class.
Takes care of the following:

- Initialize the execute- and close button
- Initialize the Progressbar
- Call function (afterSuccess, optional) after the service-call is finished to get your stored/gathered data

Used this base-class in your constructor like this:

this.addInterface("PowerToolsBase", [this]); 

It currently only supports powertools that are opened in a popup and need to show the progressbar and a execute and close button.

See the 'ImageUploader.js' (Image-uploader powertool) for an example on how to implement this baseclass.